requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
defusedxml==0.7.1
distro==1.9.0
ecdsa==0.19.2
fastapi==0.135.3
flatbuffers==25.12.19
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
magika==0.6.3
markdownify==1.2.2
markitdown==0.1.5
mpmath==1.3.0
numpy==2.4.4
onnxruntime==1.24.4
openai==2.31.0
packaging==26.1
passlib==1.7.4
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.2.0
protobuf==7.34.1
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.1
pydantic_core==2.46.1
pypdfium2==5.7.0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
requests==2.33.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.49
starlette==1.0.0
sympy==1.14.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
